media: tc358743: Skip incomplete initial frame
According to [1], TC358743XBG starts sending image data as soon as it
asked to do so. Due to this, the first frame received by V4L2 programs
is only partially filled with data. This necessitates skipping that
frame manually in userspace, which can be cumbersome.

This patch marks tc358743 as needing to have the first frame skipped.
The CSI2 receiver driver can then detect this and discard the
problematic frame before forwarding it to userspace. This is also
how the ADV7180 and other drivers handle this.

[1]: https://github.com/raspberrypi/linux/issues/4058

Suggested-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Jakub Vaněk <linuxtardis@gmail.com>